# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fazenda São Bento — Sistema de Gestão Agrícola
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /machines:
    get:
      operationId: listMachines
      tags: [machines]
      summary: List all machines
      responses:
        "200":
          description: List of machines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Machine"
    post:
      operationId: createMachine
      tags: [machines]
      summary: Create machine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMachineRequest"
      responses:
        "201":
          description: Machine created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
  /machines/{id}:
    put:
      operationId: updateMachine
      tags: [machines]
      summary: Update machine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMachineRequest"
      responses:
        "200":
          description: Machine updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
    delete:
      operationId: deleteMachine
      tags: [machines]
      summary: Delete machine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Machine deleted
...
  /harvest:
    get:
      operationId: listHarvest
      tags: [harvest]
      summary: List harvest records
      parameters:
        - name: culture
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of harvest records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HarvestRecord"
    post:
      operationId: createHarvest
      tags: [harvest]
      summary: Create harvest record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHarvestRequest"
      responses:
        "201":
          description: Harvest record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HarvestRecord"
  /harvest/{id}:
    put:
      operationId: updateHarvest
      tags: [harvest]
      summary: Update harvest record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHarvestRequest"
      responses:
        "200":
          description: Harvest record updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HarvestRecord"
    delete:
      operationId: deleteHarvest
      tags: [harvest]
      summary: Delete harvest record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Harvest record deleted
...
  /fueling:
    get:
      operationId: listFueling
      tags: [fueling]
      summary: List fueling records
      parameters:
        - name: machineId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of fueling records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FuelingRecord"
    post:
      operationId: createFueling
      tags: [fueling]
      summary: Create fueling record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFuelingRequest"
      responses:
        "201":
          description: Fueling record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FuelingRecord"
  /fueling/{id}:
    put:
      operationId: updateFueling
      tags: [fueling]
      summary: Update fueling record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFuelingRequest"
      responses:
        "200":
          description: Fueling record updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FuelingRecord"
    delete:
      operationId: deleteFueling
      tags: [fueling]
      summary: Delete fueling record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fueling record deleted
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products/inputs
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
      required:
        - message
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - email
        - password
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [admin, operador]
          default: operador
      required:
        - name
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operador]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
...
    Machine:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        model:
          type: string
        type:
          type: string
          enum: [trator, colheitadeira, caminhao, equipamento]
        location:
          type: string
        status:
          type: string
          enum: [ativo, manutencao, inativo]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - status
        - createdAt
    CreateMachineRequest:
      type: object
      properties:
        name:
          type: string
        model:
          type: string
        type:
          type: string
          enum: [trator, colheitadeira, caminhao, equipamento]
        location:
          type: string
        status:
          type: string
          enum: [ativo, manutencao, inativo]
      required:
        - name
        - type
        - status
...
    HarvestRecord:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        culture:
          type: string
          enum: [soja, feijao, milho]
        area:
          type: string
        driverName:
          type: string
        machineId:
          type: integer
        machineName:
          type: string
        quantitySacks:
          type: number
        areaHectares:
          type: number
        productivity:
          type: number
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - culture
        - area
        - driverName
        - machineId
        - quantitySacks
        - areaHectares
        - productivity
        - createdAt
    CreateHarvestRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        culture:
          type: string
          enum: [soja, feijao, milho]
        area:
          type: string
        driverName:
          type: string
        machineId:
          type: integer
        quantitySacks:
          type: number
        areaHectares:
          type: number
        notes:
          type: string
      required:
        - date
        - culture
        - area
        - driverName
        - machineId
        - quantitySacks
        - areaHectares
...
    FuelingRecord:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        machineId:
          type: integer
        machineName:
          type: string
        operatorName:
          type: string
        pump:
          type: string
        liters:
          type: number
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - machineId
        - operatorName
        - liters
        - createdAt
    CreateFuelingRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        machineId:
          type: integer
        operatorName:
          type: string
        pump:
          type: string
        liters:
          type: number
        notes:
          type: string
      required:
        - date
        - machineId
        - operatorName
        - liters
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        unit:
          type: string
          enum: [L, KG, UN, SC]
        currentStock:
          type: number
        minStock:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - unit
        - currentStock
        - createdAt
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
          enum: [L, KG, UN, SC]
        currentStock:
          type: number
        minStock:
          type: number
      required:
        - name
        - category
        - unit
        - currentStock
...