# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fazenda São Bento — Sistema de Gestão Agrícola
 * OpenAPI spec version: 0.1.0
 */
import type { ProductUnit } from "./productUnit";

export interface Product {
  id: number;
  name: string;
  category: string;
  unit: ProductUnit;
  currentStock: number;
  minStock?: number;
  createdAt: Date;
}